Started description about how to extend nightjar, which includes sample for local data store.
Improvements to the back-end implementation code, specifically around extracting out the path-based entity discovery and construction.

The documentation now includes some comments about how to extend the data store (deployment map is still a to-do).  As part of this, the code also includes a debug-only and for-example-only version of the data store which writes out to the local file system.