Add in linting to the python code.  Some shell test improvements.  Some documentation improvements.
The python code is undergoing format changes to conform to the pylint standard chosen.

The script unit tests were broken after the big move to the new layout.  They still aren't working, but at least they look at the right paths.

Updated the requirements to have the latest versions of the dependencies.